Simplified a lot of code. Started using classes.
Converted the Arm struct and ArmState struct to class with constructors,
eliminating the need for initializer functions. I moved all the servo
initialization to the constructors, which the main program will call
with the correct pin numbers.

Also made setArmState to be an instance method, which cleans stuff up a
bit.

Deleted a number of methods which we aren't using right now. We should
just make methods which we will use.

Cleaned up constants method. Made it not include the pinNumbers header
because it no longer depends on it. This forced me to include it in the
wheels code as well because it was just including the constants file
previously.

Added a tentative 5th state.